Refactor GDL90Encoder and main.cpp to improve data encoding and JSON handling. Update altitude and velocity encoding to use clamping functions for better validation. Introduce new utility functions for clamping float values to integers and normalizing degrees. Enhance JSON string writing for improved formatting and error handling in settings management.

+++ src/gdl90_encoder.cpp
@@ -118,7 +118,7 @@ uint32_t GDL90Encoder::encodeLongitude(double longitude) const {
 }
 
 uint16_t GDL90Encoder::encodeAltitude(int32_t altitude) const {
-  int32_t encoded = (altitude + 1000) / 25;
+  int64_t encoded = (static_cast<int64_t>(altitude) + 1000) / 25;
 
   if (encoded < 0) {
     encoded = 0;
